Language Token Resolution
Adding a language to a bin will now first check to see if the language token is already in the bin. If it is, it will add the contraints of the added language to the already present language instead of adding a duplicate.

Also, fixed a bug with constraints and raises an error when a language is not found in a bin.